split files, implement peer tracker
